feats

+++ FeiLib/include/FBuffer.h
@@ -19,15 +19,15 @@ public:
 
   void Append(const char* data,uint32 len);
 
-  size_t Read(Socket fd, Errno_t &errSaved);
-  size_t Write(Socket fd, int size, Errno_t &errSaved);
+  int Read(Socket fd, Errno_t &errSaved);
+  int Write(Socket fd, int size, Errno_t &errSaved);
 
   int Peek(int wantLen, char *ret); // sizeof ret >= wantlen
   void Pop(int len);
   void PopAll();
 
-  uint32 getWriteableSize() const { return m_buffer.size() - writeIdx; }
-  uint32 getReadableSize() const { return writeIdx - readIdx; }
+  int getWriteableSize() const { return int(m_buffer.size() - writeIdx); }
+  int getReadableSize() const { return int(writeIdx - readIdx); }
 
 private:
   uint32 readIdx = 0;
